Extract ctk starter into separate module
